+++ samples/Extensions/Http/HttpFunction.cs
@@ -11,19 +11,17 @@ namespace SampleApp
     public static class HttpFunction
     {
         //<docsnippet_http_trigger>
-        [Function("HttpFunction")]
+        [Function(nameof(HttpFunction))]
         public static HttpResponseData Run([HttpTrigger(AuthorizationLevel.Anonymous, "get", "post", Route = null)] HttpRequestData req,
             FunctionContext executionContext)
         {
             //<docsnippet_logging>
-            var logger = executionContext.GetLogger("HttpFunction");
+            var logger = executionContext.GetLogger(nameof(HttpFunction));
             logger.LogInformation("message logged");
             //</docsnippet_logging>
 
             var response = req.CreateResponse(HttpStatusCode.OK);
-           
             response.Headers.Add("Content-Type", "text/plain; charset=utf-8");
-
             response.WriteString("Welcome to .NET isolated worker !!");
 
             return response;