+++ sdk/Sdk.Generators/FunctionMetadataProviderGenerator/FunctionMetadataProviderGenerator.Parser.cs
@@ -42,34 +42,33 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
             /// Takes in candidate methods from the user compilation and parses them to return function metadata info as GeneratorFunctionMetadata.
             /// </summary>
             /// <param name="methods">List of candidate methods from the syntax receiver.</param>
-            public IReadOnlyList<GeneratorFunctionMetadata> GetFunctionMetadataInfo(List<MethodDeclarationSyntax> methods)
+            public IReadOnlyList<GeneratorFunctionMetadata> GetFunctionMetadataInfo(List<IMethodSymbol> methods)
             {
                 var result = ImmutableArray.CreateBuilder<GeneratorFunctionMetadata>();
 
-                var assemblyName = Compilation.Assembly.Name;
-                var scriptFile = Path.Combine(assemblyName + ".dll");
-
                 // Loop through the candidate methods (methods with any attribute associated with them)
-                foreach (MethodDeclarationSyntax method in methods)
+                foreach (IMethodSymbol method in methods)
                 {
                     CancellationToken.ThrowIfCancellationRequested();
 
-                    var model = Compilation.GetSemanticModel(method.SyntaxTree);
-
-                    if (!FunctionsUtil.IsValidFunctionMethod(_context, Compilation, model, method, out string? functionName))
+                    string? funcName = null;
+                    if (!FunctionsUtil.TryGetFunctionName(method, Compilation, out funcName))
                     {
-                        continue;
+                        _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.SymbolNotFound, Location.None, method.Name)); // would only reach here if the function attribute or method was not loaded, resulting in failure to retrieve name
                     }
+
+                    var assemblyName = method.ContainingAssembly.Name;
+                    var scriptFile = Path.Combine(assemblyName + ".dll");
 
                     var newFunction = new GeneratorFunctionMetadata
                     {
-                        Name = functionName,
-                        EntryPoint = FunctionsUtil.GetFullyQualifiedMethodName(method, model),
+                        Name = funcName,
+                        EntryPoint = FunctionsUtil.GetFullyQualifiedMethodName(method),
                         Language = Constants.Languages.DotnetIsolated,
                         ScriptFile = scriptFile
                     };
 
-                    if (!TryGetBindings(method, model, out IList<IDictionary<string, object>>? bindings, out bool hasHttpTrigger, out GeneratorRetryOptions? retryOptions))
+                    if (!TryGetBindings(method, out IList<IDictionary<string, object>>? bindings, out bool hasHttpTrigger, out GeneratorRetryOptions? retryOptions))
                     {
                         continue;
                     }
@@ -92,14 +91,14 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                 return result;
             }
 
-            private bool TryGetBindings(MethodDeclarationSyntax method, SemanticModel model, out IList<IDictionary<string, object>>? bindings, out bool hasHttpTrigger, out GeneratorRetryOptions? validatedRetryOptions)
+            private bool TryGetBindings(IMethodSymbol method, out IList<IDictionary<string, object>>? bindings, out bool hasHttpTrigger, out GeneratorRetryOptions? validatedRetryOptions)
             {
                 hasHttpTrigger = false;
                 validatedRetryOptions = null;
 
-                if (!TryGetMethodOutputBinding(method, model, out bool hasOutputBinding, out GeneratorRetryOptions? retryOptions, out IList<IDictionary<string, object>>? methodOutputBindings)
-                    || !TryGetParameterInputAndTriggerBindings(method, model, out bool supportsRetryOptions, out hasHttpTrigger, out IList<IDictionary<string, object>>? parameterInputAndTriggerBindings)
-                    || !TryGetReturnTypeBindings(method, model, hasHttpTrigger, hasOutputBinding, out IList<IDictionary<string, object>>? returnTypeBindings))
+                if (!TryGetMethodOutputBinding(method, out bool hasOutputBinding, out GeneratorRetryOptions? retryOptions, out IList<IDictionary<string, object>>? methodOutputBindings)
+                    || !TryGetParameterInputAndTriggerBindings(method, out bool supportsRetryOptions, out hasHttpTrigger, out IList<IDictionary<string, object>>? parameterInputAndTriggerBindings)
+                    || !TryGetReturnTypeBindings(method, hasHttpTrigger, hasOutputBinding, out IList<IDictionary<string, object>>? returnTypeBindings))
                 {
                     bindings = null;
                     return false;
@@ -121,7 +120,7 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                     }
                     else if (!supportsRetryOptions)
                     {
-                        _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.InvalidRetryOptions, method.GetLocation()));
+                        _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.InvalidRetryOptions, Location.None));
                         return false;
                     }
                 }
@@ -132,12 +131,9 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
             /// <summary>
             /// Checks for and returns any OutputBinding attributes associated with the method.
             /// </summary>
-            private bool TryGetMethodOutputBinding(MethodDeclarationSyntax method, SemanticModel model, out bool hasOutputBinding, out GeneratorRetryOptions? retryOptions, out IList<IDictionary<string, object>>? bindingsList)
+            private bool TryGetMethodOutputBinding(IMethodSymbol method,out bool hasOutputBinding, out GeneratorRetryOptions? retryOptions, out IList<IDictionary<string, object>>? bindingsList)
             {
-                var bindingLocation = method.Identifier.GetLocation();
-
-                var methodSymbol = model.GetDeclaredSymbol(method);
-                var attributes = methodSymbol!.GetAttributes(); // methodSymbol is not null here because it's checked in IsValidAzureFunction which is called before bindings are collected/created
+                var attributes = method!.GetAttributes(); // methodSymbol is not null here because it's checked in IsValidAzureFunction which is called before bindings are collected/created
 
                 AttributeData? outputBindingAttribute = null;
                 hasOutputBinding = false;
@@ -147,7 +143,7 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                 {
                     if (SymbolEqualityComparer.Default.Equals(attribute.AttributeClass?.BaseType, _knownFunctionMetadataTypes.RetryAttribute))
                     {
-                        if (TryGetRetryOptionsFromAttribute(attribute, method.GetLocation(), out GeneratorRetryOptions? retryOptionsFromAttr))
+                        if (TryGetRetryOptionsFromAttribute(attribute, Location.None, out GeneratorRetryOptions? retryOptionsFromAttr))
                         {
                             retryOptions = retryOptionsFromAttr;
                         }
@@ -158,7 +154,7 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                         // There can only be one output binding associated with a function. If there is more than one, we return a diagnostic error here.
                         if (hasOutputBinding)
                         {
-                            _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.MultipleBindingsGroupedTogether, bindingLocation, new object[] { "Method", method.Identifier.ToString() }));
+                            _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.MultipleBindingsGroupedTogether, Location.None, new object[] { "Method", method.Name }));
                             bindingsList = null;
                             return false;
                         }
@@ -170,7 +166,7 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
 
                 if (outputBindingAttribute != null)
                 {
-                    if (!TryCreateBindingDict(outputBindingAttribute, Constants.FunctionMetadataBindingProps.ReturnBindingName, bindingLocation, out IDictionary<string, object>? bindingDict))
+                    if (!TryCreateBindingDict(outputBindingAttribute, Constants.FunctionMetadataBindingProps.ReturnBindingName, Location.None, out IDictionary<string, object>? bindingDict))
                     {
                         bindingsList = null;
                         return false;
@@ -244,29 +240,22 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
             /// <summary>
             /// Checks for and returns input and trigger bindings found in the parameters of the Azure Function method.
             /// </summary>
-            private bool TryGetParameterInputAndTriggerBindings(MethodDeclarationSyntax method, SemanticModel model, out bool supportsRetryOptions, out bool hasHttpTrigger, out IList<IDictionary<string, object>>? bindingsList)
+            private bool TryGetParameterInputAndTriggerBindings(IMethodSymbol method, out bool supportsRetryOptions, out bool hasHttpTrigger, out IList<IDictionary<string, object>>? bindingsList)
             {
                 supportsRetryOptions = false;
                 hasHttpTrigger = false;
                 bindingsList = new List<IDictionary<string, object>>();
 
-                foreach (ParameterSyntax parameter in method.ParameterList.Parameters)
+                foreach (IParameterSymbol parameter in method.Parameters)
                 {
                     // If there's no attribute, we can assume that this parameter is not a binding
-                    if (parameter.AttributeLists.Count == 0)
+                    if (!parameter.GetAttributes().Any())
                     {
                         continue;
                     }
 
-                    if (model.GetDeclaredSymbol(parameter) is not IParameterSymbol parameterSymbol)
-                    {
-                        _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.SymbolNotFound, parameter.Identifier.GetLocation(), nameof(parameterSymbol)));
-                        bindingsList = null;
-                        return false;
-                    }
-
                     // Check to see if any of the attributes associated with this parameter is a BindingAttribute
-                    foreach (var attribute in parameterSymbol.GetAttributes())
+                    foreach (var attribute in parameter.GetAttributes())
                     {
                         if (SymbolEqualityComparer.Default.Equals(attribute.AttributeClass?.BaseType?.BaseType, _knownFunctionMetadataTypes.BindingAttribute))
                         {
@@ -276,23 +265,18 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                                 hasHttpTrigger = true;
                             }
 
-                            DataType dataType = _dataTypeParser.GetDataType(parameterSymbol.Type);
+                            DataType dataType = _dataTypeParser.GetDataType(parameter.Type);
 
                             bool cardinalityValidated = false;
-                            bool supportsDeferredBinding = false;
-
-                            if (SupportsDeferredBinding(attribute, parameterSymbol.Type.ToString()))
-                            {
-                                supportsDeferredBinding = true;
-                            }
+                            bool supportsDeferredBinding = SupportsDeferredBinding(attribute, parameter.Type.ToString());
 
                             if (_cardinalityParser.IsCardinalitySupported(attribute))
                             {
                                 DataType updatedDataType = DataType.Undefined;
 
-                                if (!_cardinalityParser.IsCardinalityValid(parameterSymbol, parameter.Type, attribute, out updatedDataType))
+                                if (!_cardinalityParser.IsCardinalityValid(parameter, attribute, out updatedDataType))
                                 {
-                                    _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.InvalidCardinality, parameter.Identifier.GetLocation(), parameterSymbol.Name));
+                                    _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.InvalidCardinality, Location.None, parameter.Name));
                                     bindingsList = null;
                                     return false;
                                 }
@@ -304,9 +288,9 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                                 cardinalityValidated = true;
                             }
 
-                            string bindingName = parameter.Identifier.ValueText;
+                            string bindingName = parameter.Name;
 
-                            if (!TryCreateBindingDict(attribute, bindingName, parameter.Identifier.GetLocation(), out IDictionary<string, object>? bindingDict, supportsDeferredBinding))
+                            if (!TryCreateBindingDict(attribute, bindingName, Location.None, out IDictionary<string, object>? bindingDict, supportsDeferredBinding))
                             {
                                 bindingsList = null;
                                 return false;
@@ -418,32 +402,35 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
             /// <summary>
             /// Checks for and returns any bindings found in the Return Type of the method
             /// </summary>
-            private bool TryGetReturnTypeBindings(MethodDeclarationSyntax method, SemanticModel model, bool hasHttpTrigger, bool hasOutputBinding, out IList<IDictionary<string, object>>? bindingsList)
+            private bool TryGetReturnTypeBindings(IMethodSymbol method, bool hasHttpTrigger, bool hasOutputBinding, out IList<IDictionary<string, object>>? bindingsList)
             {
-                TypeSyntax returnTypeSyntax = method.ReturnType;
-                ITypeSymbol? returnTypeSymbol = model.GetSymbolInfo(returnTypeSyntax).Symbol as ITypeSymbol;
+                ITypeSymbol? returnTypeSymbol = method.ReturnType;
                 bindingsList = new List<IDictionary<string, object>>();
 
                 if (returnTypeSymbol is null)
                 {
-                    _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.SymbolNotFound, returnTypeSyntax.GetLocation(), nameof(returnTypeSymbol)));
+                    _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.SymbolNotFound, Location.None, nameof(returnTypeSymbol)));
                     bindingsList = null;
                     return false;
                 }
 
                 if (!SymbolEqualityComparer.Default.Equals(returnTypeSymbol, _knownTypes.VoidType) &&
-                    !SymbolEqualityComparer.Default.Equals(returnTypeSymbol, _knownTypes.TaskType))
+                    !SymbolEqualityComparer.Default.Equals(returnTypeSymbol.OriginalDefinition, _knownTypes.TaskType))
                 {
                     // If there is a Task<T> return type, inspect T, the inner type.
-                    if (SymbolEqualityComparer.Default.Equals(returnTypeSymbol, _knownTypes.TaskOfTType))
+                    if (SymbolEqualityComparer.Default.Equals(returnTypeSymbol.OriginalDefinition, _knownTypes.TaskOfTType))
                     {
-                        GenericNameSyntax genericSyntax = (GenericNameSyntax)returnTypeSyntax;
-                        var innerTypeSyntax = genericSyntax.TypeArgumentList.Arguments.First(); // Generic task should only have one type argument
-                        returnTypeSymbol = model.GetSymbolInfo(innerTypeSyntax).Symbol as ITypeSymbol;
+                        if (returnTypeSymbol is INamedTypeSymbol namedTypeSymbol)
+                        {
+                            if (namedTypeSymbol.IsGenericType)
+                            {
+                                returnTypeSymbol = namedTypeSymbol.TypeArguments.FirstOrDefault();// Generic task should only have one type argument
+                            }
+                        }
 
                         if (returnTypeSymbol is null) // need this check here b/c return type symbol takes on a new value from the inner argument type above
                         {
-                            _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.SymbolNotFound, genericSyntax.Identifier.GetLocation(), nameof(returnTypeSymbol)));
+                            _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.SymbolNotFound, Location.None, nameof(returnTypeSymbol)));
                             bindingsList = null;
                             return false;
                         }
@@ -455,7 +442,7 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                     }
                     else
                     {
-                        if (!TryGetReturnTypePropertyBindings(returnTypeSymbol, hasHttpTrigger, hasOutputBinding, returnTypeSyntax.GetLocation(), out bindingsList))
+                        if (!TryGetReturnTypePropertyBindings(returnTypeSymbol, hasHttpTrigger, hasOutputBinding, out bindingsList))
                         {
                             bindingsList = null;
                             return false;
@@ -466,7 +453,7 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                 return true;
             }
 
-            private bool TryGetReturnTypePropertyBindings(ITypeSymbol returnTypeSymbol, bool hasHttpTrigger, bool hasOutputBinding, Location returnTypeLocation, out IList<IDictionary<string, object>>? bindingsList)
+            private bool TryGetReturnTypePropertyBindings(ITypeSymbol returnTypeSymbol, bool hasHttpTrigger, bool hasOutputBinding, out IList<IDictionary<string, object>>? bindingsList)
             {
                 var members = returnTypeSymbol.GetMembers();
                 var foundHttpOutput = false;
@@ -486,7 +473,7 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                     {
                         if (foundHttpOutput)
                         {
-                            _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.MultipleHttpResponseTypes, returnTypeLocation, new object[] { returnTypeSymbol.Name }));
+                            _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.MultipleHttpResponseTypes, Location.None, new object[] { returnTypeSymbol.Name }));
                             bindingsList = null;
                             return false;
                         }
@@ -505,7 +492,7 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                                 // validate that there's only one binding attribute per property
                                 if (foundPropertyOutputAttr)
                                 {
-                                    _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.MultipleBindingsGroupedTogether, prop.Locations.FirstOrDefault(), new object[] { "Property", prop.Name.ToString() }));
+                                    _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.MultipleBindingsGroupedTogether, Location.None, new object[] { "Property", prop.Name.ToString() }));
                                     bindingsList = null;
                                     return false;
                                 }
@@ -626,7 +613,15 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                         }
                         else
                         {
-                            attrProperties[namedArgument.Key] = namedArgument.Value.Value;
+                            if (TryParseValueByType(namedArgument.Value, out object? argValue))
+                            {
+                                attrProperties[namedArgument.Key] = argValue;
+                            }
+                            else
+                            {
+                                _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.InvalidBindingAttributeArgument, attribLocation));
+                                return false;
+                            }
                         }
                     }
                 }
@@ -657,14 +652,14 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
                 return true;
             }
 
-            private bool TryLoadConstructorArguments(AttributeData attributeData, IDictionary<string, object?> dict, Location? attribLocation)
+            private bool TryLoadConstructorArguments(AttributeData attributeData, IDictionary<string, object?> arguments, Location? attributeLocation)
             {
                 IMethodSymbol? attribMethodSymbol = attributeData.AttributeConstructor;
 
                 // Check if the attribute constructor has any parameters
                 if (attribMethodSymbol is null)
                 {
-                    _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.SymbolNotFound, attribLocation, nameof(attribMethodSymbol)));
+                    _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.SymbolNotFound, attributeLocation, nameof(attribMethodSymbol)));
                     return false;
                 }
 
@@ -677,44 +672,52 @@ namespace Microsoft.Azure.Functions.Worker.Sdk.Generators
 
                     var arg = attributeData.ConstructorArguments[i];
 
-                    switch (arg.Kind)
+                    if (TryParseValueByType(arg, out object? argValue))
                     {
-                        case TypedConstantKind.Error:
-                            break;
-
-                        case TypedConstantKind.Primitive:
-                            dict[argumentName] = arg.Value;
-                            break;
-
-                        case TypedConstantKind.Enum:
-                            var enumValue = arg.Type!.GetMembers()
-                                .FirstOrDefault(m => m is IFieldSymbol field
-                                    && field.ConstantValue is object value
-                                    && value.Equals(arg.Value));
-
-                            if (enumValue is null)
-                            {
-                                return false;
-                            }
-
-                            // we want just the enumValue symbol's name (Admin, Anonymous, Function)
-                            dict[argumentName] = enumValue.Name;
-                            break;
-
-                        case TypedConstantKind.Type:
-                            break;
-
-                        case TypedConstantKind.Array:
-                            var arrayValues = arg.Values.Select(a => a.Value?.ToString()).ToArray();
-                            dict[argumentName] = arrayValues;
-                            break;
-
-                        default:
-                            break;
+                        arguments[argumentName] = argValue;
+                    }
+                    else
+                    {
+                        _context.ReportDiagnostic(Diagnostic.Create(DiagnosticDescriptors.InvalidBindingAttributeArgument, attributeLocation));
+                        return false;
                     }
                 }
 
                 return true;
+            }
+
+            private bool TryParseValueByType(TypedConstant attributeArg, out object? argValue)
+            {
+                argValue = null;
+
+                switch (attributeArg.Kind)
+                {
+                    case TypedConstantKind.Primitive:
+                        argValue = attributeArg.Value;
+                        break;
+
+                    case TypedConstantKind.Enum:
+                        var enumValue = attributeArg.Type!.GetMembers()
+                            .FirstOrDefault(m => m is IFieldSymbol field
+                                && field.ConstantValue is object value
+                                && value.Equals(attributeArg.Value));
+
+                        if (enumValue is null)
+                        {
+                            return false;
+                        }
+
+                        // we want just the enumValue symbol's name (ex: Admin, Anonymous, Function)
+                        argValue = enumValue.Name;
+                        break;
+
+                    case TypedConstantKind.Array:
+                        var arrayValues = attributeArg.Values.Select(a => a.Value?.ToString()).ToArray();
+                        argValue = arrayValues;
+                        break;
+                }
+
+                return argValue is not null;
             }
 
             /// <summary>