+++ test/E2ETests/E2ETests/HttpEndToEndTests.cs
@@ -8,6 +8,7 @@ using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Xunit;
+using Xunit.Abstractions;
 
 namespace Microsoft.Azure.Functions.Tests.E2ETests
 {
@@ -16,9 +17,10 @@ namespace Microsoft.Azure.Functions.Tests.E2ETests
     {
         private readonly FunctionAppFixture _fixture;
 
-        public HttpEndToEndTests(FunctionAppFixture fixture)
+        public HttpEndToEndTests(FunctionAppFixture fixture, ITestOutputHelper testOutputHelper)
         {
             _fixture = fixture;
+            _fixture.TestLogs.UseTestLogger(testOutputHelper);
         }
 
         [Theory]
@@ -43,7 +45,7 @@ namespace Microsoft.Azure.Functions.Tests.E2ETests
             }
         }
 
-        [Theory]
+        [Theory(Skip = "TODO: https://github.com/Azure/azure-functions-dotnet-worker/issues/1910")]
         [InlineData("HelloFromJsonBody", "{\"Name\": \"Whitney\"}", "application/json", HttpStatusCode.OK, "Hello Whitney")]
         [InlineData("HelloFromJsonBody", "{\"Name\": \"麵🍜\"}", "application/json", HttpStatusCode.OK, "Hello 麵🍜")]
         [InlineData("HelloFromJsonBody", "{\"Name\": \"Bob\"}", "application/octet-stream", HttpStatusCode.OK, "Hello Bob")]